add-revision: build data in mutation
